moved functions into main file and config elements into their own
