+++ README.md
@@ -1,7 +1,21 @@
 # generative-art-node
 Create generative art by using the canvas api and node js
 
-## Dependencies
-`
-npm install canvas
-`
+## Installation
+
+Run the following command to install dependencies:
+
+```
+npm i
+```
+
+## Usage
+
+
+
+
+## Deployment
+
+## Changelog
+
+## Contributing
