+++ index.js
@@ -1,97 +1,11 @@
-const fs = require("fs");
 const myArgs = process.argv.slice(2);
-const { createCanvas, loadImage } = require("canvas");
-const { layers, width, height } = require("./src/config.js");
-const console = require("console");
-
-const canvas = createCanvas(width, height);
-const ctx = canvas.getContext("2d");
-const edition = myArgs.length > 0 ? Number(myArgs[0]) : 1;
-let metadata = [];
-let attributes = [];
-let hash = [];
-let decodedHash = [];
-
-const buildDir = __dirname + '/build';
-const metDataFile = '_metadata.json';
-
-const buildSetup = () => {
-  if (fs.existsSync(buildDir)) {
-    fs.rmdirSync(buildDir, { recursive: true });
-  }
-  fs.mkdirSync(buildDir);
-}
-
-const saveLayer = (_canvas, _edition) => {
-  fs.writeFileSync(`${buildDir}/${_edition}.png`, _canvas.toBuffer("image/png"));
-};
-
-const addMetadata = (_edition) => {
-  let dateTime = Date.now();
-  let tempMetadata = {
-    hash: hash.join(""),
-    decodedHash: decodedHash,
-    edition: _edition,
-    date: dateTime,
-    attributes: attributes,
-  };
-  metadata.push(tempMetadata);
-  attributes = [];
-  hash = [];
-  decodedHash = [];
-};
-
-const addAttributes = (_element, _layer) => {
-  let tempAttr = {
-    id: _element.id,
-    layer: _layer.name,
-    name: _element.name,
-    rarity: _element.rarity,
-  };
-  attributes.push(tempAttr);
-  hash.push(_layer.id);
-  hash.push(_element.id);
-  decodedHash.push({ [_layer.id]: _element.id });
-};
-
-const drawLayer = async (_layer, _edition) => {
-  let element =
-    _layer.elements[Math.floor(Math.random() * _layer.elements.length)];
-  addAttributes(element, _layer);
-  const image = await loadImage(`${_layer.location}${element.fileName}`);
-  ctx.drawImage(
-    image,
-    _layer.position.x,
-    _layer.position.y,
-    _layer.size.width,
-    _layer.size.height
-  );
-  saveLayer(canvas, _edition);
-};
-
-const createFiles = () => {
-  for (let i = 1; i <= edition; i++) {
-    layers.forEach((layer) => {
-      drawLayer(layer, i);
-    });
-    addMetadata(i);
-    console.log("Creating edition " + i);
-  }
-}
-
-const createMetaData = () => {
-  fs.stat(`${buildDir}/${metDataFile}`, (err, stat) => {
-    if(err == null || err.code === 'ENOENT') {
-      fs.writeFileSync(`${buildDir}/${metDataFile}`, JSON.stringify(metadata));
-    } else {
-        console.log('Oh no, error: ', err.code);
-    }
-  });
-}
+const { buildSetup, createFiles, createMetaData } = require("./src/main.js");
+const { defaultEditions } = require("./src/config.js");
+const edition = myArgs.length > 0 ? Number(myArgs[0]) : defaultEditions;
 
 (() => {
   buildSetup();
-  createFiles();
+  createFiles(edition);
   createMetaData();
 })();
